extract factory functions to separate file
