+++ pkg/resources/interface.go
@@ -18,9 +18,6 @@ package resources
 
 import (
 	corev1 "k8s.io/api/core/v1"
-	resourcehelpers "k8s.io/component-helpers/resource"
-
-	"sigs.k8s.io/kueue/pkg/features"
 )
 
 // Requests represents an abstract collection of resource requests.
@@ -35,45 +32,4 @@ type Requests interface {
 	ForEach(fn func(name corev1.ResourceName, val int64))
 	Len() int
 	IsEmpty() bool
-}
-
-// CreateEmpty creates an empty Requests instance based on feature gates.
-func CreateEmpty() Requests {
-	if features.Enabled(features.VectorizedResourceRequests) {
-		return &SliceRequests{}
-	}
-	return MapRequests{}
-}
-
-// NewRequestsFromMap creates a Requests instance from a MapRequests map based on feature gates.
-func NewRequestsFromMap(m MapRequests) Requests {
-	if len(m) == 0 {
-		return nil
-	}
-	if features.Enabled(features.VectorizedResourceRequests) {
-		sr := toSliceRequests(m)
-		return &sr
-	}
-	return m
-}
-
-// NewRequestsFromResourceList creates a Requests instance from a corev1.ResourceList based on feature gates.
-func NewRequestsFromResourceList(rl corev1.ResourceList) Requests {
-	if len(rl) == 0 {
-		return nil
-	}
-	if features.Enabled(features.VectorizedResourceRequests) {
-		sr := ResourceListToSliceRequests(rl)
-		return &sr
-	}
-	return NewMapRequests(rl)
-}
-
-// NewRequestsFromPodSpec creates a Requests instance from a PodSpec based on feature gates.
-func NewRequestsFromPodSpec(podSpec *corev1.PodSpec) Requests {
-	if podSpec == nil {
-		return nil
-	}
-	rl := resourcehelpers.PodRequests(&corev1.Pod{Spec: *podSpec}, resourcehelpers.PodResourcesOptions{})
-	return NewRequestsFromResourceList(rl)
 }
